fix incorrect 'bad document error' when searching for 'TVAR'
I suspect that when the eLink returns the same number of results as the initial esearch, we do not get a new query key. The missing query_key in the link result caused the "Bad Document" error. We no longer assume we will get a query key back and use the previous query from the esearch if we don't get one from eLink.